Fix timeseries in redfish parser
While trying to enable the redfish job in the CI we found:
`Duplicated timeseries in CollectorRegistry: {'baremetal_fan_status'}`,
this is due to the fact that  we were creating multiple `Gauge` for
the metric and Prometheus does not support this.

We should pre-create the Gauge and update with the different sets of
labels and values that were collected.

Change-Id: I4f4ea8e580ccb439afa7c8e14fabb94c2d124b17