Prometheus-exporter to use DevStack's neutron"-legacy" module
In the last PTG the Neutron team discussed and decided to undeprecate
the neutron-legacy module in DevStack because that's the module being
used (almost) everywhere and it works. The lib/neutron was an attempt
to refactor the old module but, in the last few years it hasn't gained
any traction and due to the lack of features and people to work on it,
it's going to be removed from DevStack eventually.

Below is a snippet from the PTG summary email [0] about this topic:

<snippet>
In Devstack there are currently 2 modules which can configure
Neutron. Old one called "lib/neutron-legacy" and the new one called
"lib/neutron". It is like that since many cycles that
"lib/neutron-legacy"
is deprecated. But it is still used everywhwere. New module isn't still
finished and isn't working fine.  This is very confusing for users as
really maintained and recommended is still "lib/neutron-legacy" module.

During the discussion Sean Collins explained us that originally this
new module was created as an attempt to refactor old module, and to
make Neutron in the Devstack better to maintain. But now we see that
this process failed as new module isn't still used and we don't have
any cycles to work on it. So our final conclusion is to "undeprecate"
old "lib/neutron-legacy" and get rid of the new module.
</snippet>

This patch changes the ironic-prometheus-exporter jobs to use the old
Neutron module in DevStack.

[0]
http://lists.openstack.org/pipermail/openstack-discuss/2020-June/015368.html
[1]
http://codesearch.openstack.org/?q=neutron-api%3A%20true&i=nope&files=&repos=

Change-Id: Ie8f795e8ff13461e14ff8b9e5bcf3b0ed8b629f5
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>