devstack: fix plugin for local usage
The wsgi webserver used by the test, which is *not* a hard
requirement for the software's usage, is gunicorn, which is
not present by default unless something has already
installed it. It is in OpenStack's requirements, so it is
not impossible.

Regardless, install so the devstack plugin "just works"
when run locally on a clean/new VM.

Change-Id: I8a379aa7ca94c27211aebee77dfefc742c3f1223