Trivial: Fix TypeError from None value being returned in sensor data
When prometheus data is being transformed, there is a risk that a
None could be introduced into the data stream whcih then creates an
exception because you cannot float(None). None is also invalid data
in that we shouldn't even try to record anything.

So, this change just adds a check to exclude from trying to record
a None value. This will result in missing data if a sensor
turned from 0 to None somehow, but realistically the sensor endpoint
is yielding None to begin with in that case through a missing value
anyhow.

Change-Id: I7a7197228e5102c68e580e87fa8f0485ff2eb074